# server-application/restapi/src/main/resources/openapi-specification.yml
openapi: 3.0.3
info:
  title: Mobile NotWork Game REST API
  version: "1.0"
...
paths:
  /spieler:
    get:
      summary: Gibt einen Spieler zu gegebener sid zurueck oder listet alle Spieler auf.
      description: Gibt einen Spieler zu gegebener sid zurueck oder listet alle Spieler auf.
      tags:
        - SpielerRESTController
      security:
        - spielerTokenAuth: []
      operationId: getSpieler
      parameters:
        - name: sid
          in: query
          description: Die Spieler ID
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Erfolgreiche Ermittlung der Spieler
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Spieler'
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /spieler/benutzername/{sid}:
    get:
      summary: Benutzernamen fuer Spieler ermitteln
      description: Fuer eine gegebene sid den zugehoerigen Benutzernamen ermitteln
      operationId: getBenutzername
      tags:
        - SpielerRESTController
      parameters:
        - name: sid
          in: path
          description: Die ID des Spielers.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Benutzername erfolgreich ermittelt
          content:
            application/json:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /karte/get-ksid-fuer-standort:
    get:
      summary: ermittelt fuer gegebene Koordinaten die Kartensegment ID
      description: ermittelt fuer gegebene Koordinaten die Kartensegment ID
      operationId: getKsidFuerStandort
      tags:
        - KarteRESTController
      security:
        - spielerTokenAuth: []

      parameters:
        - name: Lat
          in: query
          description: Latitude des Standorts
          required: true
          schema:
            type: number
        - name: Lon
          in: query
          description: Longitude des Standorts
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Kartensegment wurde ermittelt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KsidFuerStandort'
        '400':
          description: Ungueltige Anfrage
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
  /karte/get-kartenbereich-daten/spieler-ansicht:
    get:
      summary: Kartensegmente inkl. Messungen fuer gegebenen Bereich fuer die Sicht des Spielers ermitteln
      description: Ermitteln der Kartensegmente inkl. Messungen fuer die Kartenansicht in einem bestimmten Bereich fuer die Ansicht des Spielers.
      operationId: getKartenbereichDatenFuerSpieler
      tags:
        - KarteRESTController
      security:
        - spielerTokenAuth: []

      parameters:
        - name: linksObenLat
          in: query
          description: Latitude des linken oberen Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: linksObenLon
          in: query
          description: Longitude des linken oberen Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: rechtsObenLat
          in: query
          description: Latitude des rechten oberen Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: rechtsObenLon
          in: query
          description: Longitude des rechten oberen Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: rechtsUntenLat
          in: query
          description: Latitude des rechten unteren Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: rechtsUntenLon
          in: query
          description: Longitude des rechten unteren Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: linksUntenLat
          in: query
          description: Latitude des linken unteren Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: linksUntenLon
          in: query
          description: Longitude des linken unteren Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: filter
          in: query
          description: Filter fuer die Karte
          required: true
          schema:
            type: string
        - name: filterValue
          in: query
          description: FilterValue fuer die Karte
          required: true
          schema:
            type: integer
        - name: sid
          in: query
          description: ID des Spielers
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Messungen und Kartensegmente erfolgreich ermittelt
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KartenbereichDatenfuerSpieler'
        '400':
          description: Ungueltige Anfrage
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
  /karte/get-kartenbereich-daten/admin-ansicht:
    get:
      summary: Kartensegmente inkl. Messungen fuer gegebenen Bereich fuer die Sicht des Administrators ermitteln
      description: Ermitteln der Kartensegmente inkl. Messungen fuer die Kartenansicht in einem bestimmten Bereich fuer die Ansicht des Administrators.
      operationId: getKartenbereichDatenFuerAdmin
      tags:
        - KarteRESTController
      security:
        - adminTokenAuth: []

      parameters:
        - name: linksObenLat
          in: query
          description: Latitude des linken oberen Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: linksObenLon
          in: query
          description: Longitude des linken oberen Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: rechtsObenLat
          in: query
          description: Latitude des rechten oberen Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: rechtsObenLon
          in: query
          description: Longitude des rechten oberen Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: rechtsUntenLat
          in: query
          description: Latitude des rechten unteren Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: rechtsUntenLon
          in: query
          description: Longitude des rechten unteren Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: linksUntenLat
          in: query
          description: Latitude des linken unteren Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: linksUntenLon
          in: query
          description: Longitude des linken unteren Punktes des anzuzeigenden Bereichs.
          required: true
          schema:
            type: number
        - name: filter
          in: query
          description: Filter fuer die Karte
          required: true
          schema:
            type: string
        - name: filterValue
          in: query
          description: FilterValue fuer die Karte
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Messungen und Kartensegmente erfolgreich ermittelt
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KartenbereichDatenfuerAdmin'
        '400':
          description: Ungueltige Anfrage
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
...
    Spieler:
      type: object
      properties:
        sid:
          type: integer
          format: int32
        benutzername:
          type: string
        geraeteid:
          type: string
        einwilligungagb:
          type: boolean
...
    MapHexagonCorner:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
    MapHexagon:
      type: object
      properties:
        punkt1:
          $ref: '#/components/schemas/MapHexagonCorner'
        punkt2:
          $ref: '#/components/schemas/MapHexagonCorner'
        punkt3:
          $ref: '#/components/schemas/MapHexagonCorner'
        punkt4:
          $ref: '#/components/schemas/MapHexagonCorner'
        punkt5:
          $ref: '#/components/schemas/MapHexagonCorner'
        punkt6:
          $ref: '#/components/schemas/MapHexagonCorner'
    KartenbereichDatenfuerSpieler:
      type: object
      properties:
        ksid:
          type: integer
        punkte:
          $ref: '#/components/schemas/MapHexagon'
        mobilfunkstandardAsString:
          type: string
        mobilfunkanbieterAsString:
          type: string
        rssi:
          type: integer
        rtt:
          type: integer
        snr:
          type: integer
        erreichtePunkte:
          type: integer
    KartenbereichDatenfuerAdmin:
      type: object
      properties:
        ksid:
          type: integer
          format: int32
        punkte:
          $ref: '#/components/schemas/MapHexagon'
        mobilfunkstandardAsString:
          type: string
        mobilfunkanbieterAsString:
          type: string
        rssi:
          type: integer
          format: int32
        rtt:
          type: integer
          format: int32
        snr:
          type: integer
          format: int32
        erreichtePunkte:
          type: integer
          format: int32
        spielerBenutzername:
          type: string
        sid:
          type: integer
          format: int32
    KsidFuerStandort:
      type: object
      properties:
        ksid:
          type: integer
          format: int32
  securitySchemes:
    spielerTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            type: string
            example: "Angeforderte Ressource wurde nicht gefunden."
    NotAuthenticated:
      description: Nicht authentifiziert
      content:
        application/json:
          schema:
            type: string
            example: "Nicht authentifiziert, bitte melden Sie sich an."